+++ src/direction/impls.rs
@@ -1,4 +1,4 @@
-use std::ops::Neg;
+use std::ops::{Add, Neg, Sub};
 
 use crate::{DiagonalDirection, Direction};
 
@@ -15,5 +15,37 @@ impl Neg for Direction {
 
     fn neg(self) -> Self::Output {
         self.const_neg()
+    }
+}
+
+impl Add<usize> for Direction {
+    type Output = Self;
+
+    fn add(self, rhs: usize) -> Self::Output {
+        self.rotate_right(rhs)
+    }
+}
+
+impl Add<usize> for DiagonalDirection {
+    type Output = Self;
+
+    fn add(self, rhs: usize) -> Self::Output {
+        self.rotate_right(rhs)
+    }
+}
+
+impl Sub<usize> for Direction {
+    type Output = Self;
+
+    fn sub(self, rhs: usize) -> Self::Output {
+        self.rotate_left(rhs)
+    }
+}
+
+impl Sub<usize> for DiagonalDirection {
+    type Output = Self;
+
+    fn sub(self, rhs: usize) -> Self::Output {
+        self.rotate_left(rhs)
     }
 }